+++ src/lib/components/dialog/datasource/DatasourceDialog.svelte
@@ -2,13 +2,14 @@
 	import * as Dialog from '$lib/components/ui/dialog/index.js';
 	import * as ToggleGroup from '$lib/components/ui/toggle-group';
 	import * as Card from '$lib/components/ui/card';
+	import { Button } from '$lib/components/ui/button';
+	import { Input } from '$lib/components/ui/input';
 	import { Label } from '$lib/components/ui/label/index.js';
-	import type { DatasourceConfig } from '$lib/lens/types';
-	import { Button } from '../ui/button';
+	import type { DatasourceConfig, StoreType } from '$lib/lens/types';
 	import Icon from '@iconify/svelte';
-	import { Input } from '../ui/input';
 	import type { Component } from 'svelte';
 	import AmazonS3Options from './AmazonS3Options.svelte';
+	import GoogleCloudStorageOptions from './GoogleCloudStorageOptions.svelte';
 
 	export function show(): Promise<DatasourceConfig> {
 		return new Promise<DatasourceConfig>((accept, reject) => {
@@ -18,20 +19,17 @@
 		});
 	}
 
-	type DataSource = 's3' | 'gcs';
-
 	type OptionsComponent = Component<{}, { getConfig: () => any }>;
 	type DatasourceItem = {
 		label: string;
 		icon: string;
+		defaultUrl: string;
 		get url(): string;
 		set url(value: string);
 		options?: OptionsComponent;
 	};
 
-	let options = $state<OptionsComponent | undefined>(undefined);
-
-	let datasources: Record<DataSource, DatasourceItem> = {
+	let datasources: Record<StoreType, DatasourceItem> = {
 		s3: useDatasource({
 			label: 'Amazon S3',
 			icon: 'mdi:aws',
@@ -41,13 +39,14 @@
 		gcs: useDatasource({
 			label: 'Google Cloud Storage',
 			icon: 'mdi:google',
-			defaultUrl: 'gcp://',
-			options: undefined
+			defaultUrl: 'gs://',
+			options: GoogleCloudStorageOptions
 		})
 	};
 
 	let open = $state(false);
-	let selected: DataSource = $state('s3');
+	let selected = $state<StoreType>('s3');
+	let options: OptionsComponent;
 
 	let accept_: (config: DatasourceConfig) => void;
 	let reject_: () => void;
@@ -68,6 +67,7 @@
 		return {
 			label,
 			icon,
+			defaultUrl,
 			get url() {
 				return url;
 			},
@@ -81,12 +81,14 @@
 	function closeDialog() {
 		open = false;
 		const datasource = datasources[selected];
-		let storeConfig: any = {};
-		storeConfig[selected] = options?.getConfig();
+		const storeConfig = options?.getConfig();
 
 		const config: DatasourceConfig = {
 			url: datasource.url,
-			store: storeConfig
+			store: {
+				kind: selected,
+				config: storeConfig
+			}
 		};
 
 		if (accept_) accept_(config);
@@ -125,8 +127,7 @@
 				<Input bind:value={datasources[selected].url} />
 
 				<p class="text-sm text-muted-foreground">
-					The URI your datasource will be accessible from. You can provide a prefix or bucket (eg
-					{datasources[selected].url}my-bucket)
+					The URI your datasource will be accessible from. You can provide a prefix or bucket
 				</p>
 			</div>
 			<Label>Options</Label>